requirements.txt
opencv-python==4.5.3.56
pillow
pygame

